stylized authentication page, put css variables in helpers & import them in app.css
