+++ src/client/app.css
@@ -1,0 +1,14 @@
+@import url("https://fonts.googleapis.com/css?family=Orbitron&display=swap");
+@import url("https://fonts.googleapis.com/css?family=Khand|Orbitron&display=swap");
+
+@import "./helpers/variables.css";
+
+* {
+  box-sizing: border-box;
+  margin: 0;
+  padding: 0;
+}
+
+body {
+    height: 100%;
+}
