+++ src/client/app.js
@@ -3,6 +3,7 @@ import ReactDOM from 'react-dom';
 import { Provider } from 'react-redux';
 import store from './store/store';
 import Routers from './routers';
+import './app.css';
 
 const app = (
   <Provider store={store}>
