+++ src/client/components/formField/formField.css
@@ -5,10 +5,15 @@
 .field input {
   width: 100%;
   height: 1rem;
-  margin-bottom: 1rem;
+  margin: 0.5rem 0 1rem;
   padding-left: 0.5rem;
   border: none;
-  border-bottom: 2px solid rgb(20, 19, 17);
-  outline: none;
-  background-color: transparent;
+  background-color: var(--authInputBackground-blue-superlight);
+}
+
+@media screen and (max-width: 500px) {
+  .field input {
+    height: 0.8rem;
+    margin: 0 0 0.5rem;
+  }
 }
