+++ src/client/containers/main/main.css
@@ -1,15 +1,4 @@
-@import url("https://fonts.googleapis.com/css?family=Orbitron&display=swap");
-@import url("https://fonts.googleapis.com/css?family=Khand|Orbitron&display=swap");
-
-@import "./variables.css";
-
-* {
-  box-sizing: border-box;
-  margin: 0;
-  padding: 0;
-}
-
-body {
+.main {
   font-family: var(--font-ordinary);
   color: var(--mainText-white);
   background: var(--bodyBackground-dark-light);
