--- src/client/index.js
@@ -1,5 +1,0 @@
-import React from 'react';
-import ReactDOM from 'react-dom';
-import App from './app';
-
-ReactDOM.render(<App />, document.getElementById('root'));
