+++ src/client/routers.js
@@ -3,12 +3,12 @@ import { Route, Switch } from 'react-router';
 import { BrowserRouter } from 'react-router-dom';
 
 import Main from './containers/main/main';
-import Authentification from './containers/authentification/authentification';
+import Authentication from './containers/authentication/authentication';
 
 const Routers = () => (
   <BrowserRouter>
     <Switch>
-      <Route path="/" exact component={Authentification} />
+      <Route path="/" exact component={Authentication} />
       <Route path="/main" component={Main} />
     </Switch>
   </BrowserRouter>
